GPUProfiler의 elapsed time(delta time)의 자료형을 double로 변경하여, 더욱 자세한 퍼포먼스를 측정할 수 있게 하였습니다

+++ Sources/Runtime/Rendering/GPUProfiler.cpp
@@ -94,7 +94,7 @@ namespace Mile
                data.ElapsedTimeHz = endTime - beginTime;
 
                double frequency = static_cast<double>(disjointData.Frequency);
-               data.ElapsedTime = static_cast<UINT64>((data.ElapsedTimeHz / frequency) * 1000.0);
+               data.ElapsedTime = ((data.ElapsedTimeHz / frequency) * 1000.0);
             }
 
             m_profileTimes[(*dataItr).first] = data.ElapsedTime;
